+++ pkgs/common/userChromeJS.nix
@@ -1,6 +1,5 @@
 {
   runCommand,
-  p7zip,
   lib,
   source,
 }:
@@ -10,12 +9,11 @@ runCommand "firefox-utils" {
     if (builtins.hasAttr "date" source)
     then source.date
     else lib.removePrefix "v" source.version;
-  nativeBuildInputs = [p7zip];
   meta = with lib; {
-    homepage = "https://github.com/xiaoxiaoflood/firefox-scripts";
+    homepage = "https://github.com/benzBrake/userChrome.js-Loader";
     description = "Firefox scripts ";
   };
 } ''
   mkdir -p $out/share
-  7z x -y $src -o$out/share
+  cp -rv $src/profile/chrome/* $out/
 ''