+++ pkgs/darwin/firefox/default.nix
@@ -3,12 +3,11 @@
   sources,
   withEsr ? true,
   withFx ? true,
-  p7zip,
   lib,
   fetchurl,
   ...
 }: let
-  fxSrc = sources.fx-folder.src;
+  fxSrc = sources.userChromeJS.src;
   pname =
     if withEsr
     then "firefox-esr"
@@ -22,18 +21,13 @@ in
   mkDarwinApp rec {
     inherit pname version src;
     appname = "Firefox";
-    nativeBuildInputs = [p7zip];
     postInstall =
       ''
       ''
       + lib.optionalString withFx ''
-        mkdir -p $out/tmp
-        7z x ${fxSrc} -o$out/tmp
-        ls $out/tmp
-        cp -ar $out/tmp/config.js $out/Applications/Firefox.app/Contents/Resources/config.js
+        cp -ar ${fxSrc}/program/config.js $out/Applications/Firefox.app/Contents/Resources/config.js
         mkdir -p $out/Applications/Firefox.app/Contents/Resources/defaults/pref
-        cp -ar $out/tmp/defaults/pref/config-prefs.js $out/Applications/Firefox.app/Contents/Resources/defaults/pref
-        rm -rf $out/tmp
+        cp -ar ${fxSrc}/defaults/pref/config-prefs.js $out/Applications/Firefox.app/Contents/Resources/defaults/pref
       '';
     meta = {
       description = "Mozilla Firefox, free web browser (binary package)";